requirements.txt
Django==1.11
mysqlclient==1.3.10
wheel==0.29.0
gunicorn==19.7.1
psycopg2==2.7.1
google-api-python-client
